+++ deploy/servctl_utils/deploy_command_utils.py
@@ -107,8 +107,6 @@ def deploy_secrets(settings, components=None):
 def deploy_elasticsearch(settings):
     print_separator("elasticsearch")
 
-    docker_build("elasticsearch", settings, ["--build-arg ELASTICSEARCH_SERVICE_PORT=%s" % settings["ELASTICSEARCH_SERVICE_PORT"]])
-
     if settings["ONLY_PUSH_TO_REGISTRY"]:
         return
 
@@ -149,8 +147,8 @@ def deploy_elasticsearch(settings):
 def _set_elasticsearch_kubernetes_resources():
     has_kube_resource = run('kubectl explain elasticsearch', errors_to_ignore=["server doesn't have a resource type", "couldn't find resource for"])
     if not has_kube_resource:
-        run('kubectl create -f https://download.elastic.co/downloads/eck/1.8.0/crds.yaml')
-        run('kubectl apply -f https://download.elastic.co/downloads/eck/1.8.0/operator.yaml')
+        run('kubectl create -f https://download.elastic.co/downloads/eck/1.9.1/crds.yaml')
+        run('kubectl apply -f https://download.elastic.co/downloads/eck/1.9.1/operator.yaml')
 
 
 def deploy_elasticsearch_snapshot_config(settings):