+++ seqr/views/apis/variant_search_api_tests.py
@@ -406,6 +406,7 @@ class VariantSearchAPITest(object):
         self.assertSetEqual(set(response_json), response_keys)
         self.assertEqual(len(response_json['savedSearchesByGuid']), 3)
         self.assertTrue(PROJECT_GUID in response_json['projectsByGuid'])
+        self.assertTrue(response_json['projectsByGuid'][PROJECT_GUID]['searchContextLoaded'])
         self.assertTrue('F000001_1' in response_json['familiesByGuid'])
         self.assertTrue('AG0000183_test_group' in response_json['analysisGroupsByGuid'])
 
@@ -534,7 +535,7 @@ class VariantSearchAPITest(object):
         self.assertEqual(len(saved_searches), 1)
         search_guid = next(iter(saved_searches))
         self.assertDictEqual(saved_searches[search_guid], {
-            'savedSearchGuid': search_guid, 'name': 'Test Search', 'search': SEARCH, 'createdById': 13,
+            'savedSearchGuid': search_guid, 'name': 'Test Search', 'search': SEARCH, 'createdById': 13, 'order': None,
         })
 
         # Test no errors if duplicate searches get created
@@ -566,7 +567,7 @@ class VariantSearchAPITest(object):
         response = self.client.post(update_saved_search_url, content_type='application/json', data=json.dumps(body))
         self.assertEqual(response.status_code, 200)
         self.assertDictEqual(response.json()['savedSearchesByGuid'][search_guid], {
-            'savedSearchGuid': search_guid, 'name': 'Updated Test Search', 'search': SEARCH, 'createdById': 13,
+            'savedSearchGuid': search_guid, 'name': 'Updated Test Search', 'search': SEARCH, 'createdById': 13, 'order': None,
         })
 
         delete_saved_search_url = reverse(delete_saved_search_handler, args=[search_guid])