+++ ui/pages/Search/components/filters/LocusListSelector.jsx
@@ -1,7 +1,6 @@
-import React from 'react'
 import PropTypes from 'prop-types'
+import React from 'react'
 import { connect } from 'react-redux'
-
 import { Dropdown } from 'shared/components/form/Inputs'
 import { LocusListItemsLoader } from 'shared/components/LocusListLoader'
 import { getSearchedProjectsLocusListOptions } from '../../selectors'
@@ -24,8 +23,23 @@ class BaseLocusListDropdown extends React.Component {
 
   componentDidUpdate(prevProps) {
     const { locusList, onChange } = this.props
+
     if (prevProps.locusList.rawItems !== locusList.rawItems) {
-      const { locusListGuid, rawItems } = locusList
+      const { locusListGuid } = locusList
+      let { rawItems } = locusList
+
+      if (locusList.paLocusList) {
+        const CONFIDENCE_COLORS = { 0: 'none', 1: 'red', 2: 'amber', 3: 'green', 4: 'green' }
+        rawItems = locusList.items?.reduce((acc, item) => {
+          const color = CONFIDENCE_COLORS[item.pagene?.confidenceLevel || 0]
+          if (color in acc) {
+            acc[color] = [acc[color], item.display].filter(val => val).join(', ')
+          }
+
+          return acc
+        }, { green: '', amber: '', red: '', none: '' })
+      }
+
       onChange({ locusListGuid, rawItems })
     }
   }