+++ ui/shared/components/panel/variants/VariantIndividuals.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import { connect } from 'react-redux'
-import { Popup, Icon, Header, Divider } from 'semantic-ui-react'
+import { Popup, Icon, Header, Divider, Label } from 'semantic-ui-react'
 
 import { getSortedIndividualsByFamily, getGenesById } from 'redux/selectors'
 import PedigreeIcon from '../../icons/PedigreeIcon'
@@ -45,6 +45,7 @@ const IndividualCell = styled.div`
   
   .ui.header {
     padding-top: 3px;
+    margin-bottom: 3px;
   }
 `
 
@@ -84,6 +85,24 @@ const missingParentVariant = variant => (parentGuid) => {
 const isHemiUPDVariant = (numAlt, variant, individual) => (
   numAlt === 2 && [individual.maternalGuid, individual.paternalGuid].some(missingParentVariant(variant)))
 
+const isCalled = val => Number.isInteger(val) && val >= 0
+
+const getGentoypeCnWarning = (genotype, svType, isHemiX) => {
+  const refCn = isHemiX ? 1 : 2
+  const hasGentotype = isCalled(genotype.numAlt)
+
+  if ((svType === 'DUP' && genotype.cn < refCn) || (svType === 'DEL' && genotype.cn > refCn)) {
+    return `Copy Number does not match Call Type. Copy number calling may be unreliable for small events${hasGentotype ? ', however genotype call is likely accurate' : ''}`
+  }
+
+  if (hasGentotype && (
+    (genotype.numAlt === 0 && genotype.cn !== refCn) || (genotype.numAlt > 0 && genotype.cn === refCn))) {
+    return 'Copy number does not match genotype. Copy number calling may be unreliable for small events, however genotype call is likely accurate'
+  }
+
+  return null
+}
+
 const Allele = styled.div.attrs(({ isAlt, variant }) => ({ children: isAlt ? variant.alt : variant.ref }))`
   display: inline-block;
   max-width: 50px;
@@ -102,28 +121,24 @@ Allele.propTypes = {
   variant: PropTypes.object,
 }
 
+const copyNumberGenotype = (cn, isHemiX) => (
+  <span>
+    CN: &nbsp;
+    {cn !== (isHemiX ? 1 : 2) ? <b><i>{cn}</i></b> : cn}
+  </span>
+)
+
 const svGenotype = (genotype, isHemiX) => {
-  const hasGenotype = Number.isInteger(genotype.numAlt) && genotype.numAlt >= 0
-  const isAltCn = Number.isInteger(genotype.cn) && genotype.cn !== (isHemiX ? 1 : 2)
-  if (!hasGenotype) {
-    return (
-      <span>
-        CN: &nbsp;
-        {isAltCn ? <b><i>{genotype.cn}</i></b> : genotype.cn}
-      </span>
-    )
+  const cnDisplay = isCalled(genotype.cn) && copyNumberGenotype(genotype.cn, isHemiX)
+  if (!isCalled(genotype.numAlt)) {
+    return cnDisplay
   }
   return (
     <span>
       {/* eslint-disable-next-line react/jsx-one-expression-per-line */}
       {isHemiX || genotype.numAlt < 2 ? 'ref' : <b><i>alt</i></b>}/{genotype.numAlt > 0 ? <b><i>alt</i></b> : 'ref'}
-      {isAltCn && (
-        <span>
-          <br />
-          CN: &nbsp;
-          <b><i>{genotype.cn}</i></b>
-        </span>
-      )}
+      {cnDisplay && <br />}
+      {cnDisplay}
     </span>
   )
 }
@@ -132,7 +147,7 @@ export const Alleles = React.memo(({ genotype, variant, isHemiX, warning }) => (
   <AlleleContainer>
     {warning && (
       <Popup
-        flowing
+        wide
         trigger={<Icon name="warning sign" color="yellow" />}
         content={
           <div>
@@ -226,10 +241,8 @@ const Genotype = React.memo(({ variant, individual, isCompoundHet, genesById }) 
     return null
   }
 
-  const isNoCall = variant.svType ?
-    (!Number.isInteger(genotype.cn) && (!Number.isInteger(genotype.numAlt) || genotype.numAlt < 0)) :
-    genotype.numAlt < 0
-  if (isNoCall) {
+  const hasCnCall = isCalled(genotype.cn)
+  if (!hasCnCall && !isCalled(genotype.numAlt)) {
     return <b>NO CALL</b>
   }
 
@@ -244,8 +257,20 @@ const Genotype = React.memo(({ variant, individual, isCompoundHet, genesById }) 
     warning = 'Variant absent in parents'
   }
 
-  if ((variant.svType === 'DUP' && genotype.cn < 2) || (variant.svType === 'DEL' && genotype.cn > 2)) {
-    warning = [warning, 'Copy Number does not match Call Type.'].join(warning ? '. ' : '')
+  if (hasCnCall) {
+    const cnWarning = getGentoypeCnWarning(genotype, variant.svType, isHemiX)
+    if (cnWarning) {
+      warning = warning ? `${warning}. ${cnWarning}` : cnWarning
+    }
+  }
+
+  let previousCall
+  if (genotype.newCall) {
+    previousCall = { content: 'New Call', hover: 'No overlap in previous callset', color: 'green' }
+  } else if (genotype.prevCall) {
+    previousCall = { content: 'Identical Call', hover: 'Identical call in previous callset', color: 'blue' }
+  } else if (genotype.prevOverlap) {
+    previousCall = { content: 'Overlapping Call', hover: 'Overlapping call in previous callset', color: 'teal' }
   }
 
   const hasConflictingNumAlt = genotype.otherSample && genotype.otherSample.numAlt !== genotype.numAlt
@@ -272,7 +297,12 @@ const Genotype = React.memo(({ variant, individual, isCompoundHet, genesById }) 
         />
       )}
       <Alleles genotype={genotype} variant={variant} isHemiX={isHemiX} warning={warning} />
-      <VerticalSpacer height={2} />
+      {previousCall && (
+        <Popup
+          content={previousCall.hover}
+          trigger={<Label horizontal size="mini" content={previousCall.content} color={previousCall.color} />}
+        />
+      )}
       {`${genotype.gq || genotype.qs || '-'}${variant.svType ? '' : genotype.numAlt >= 0 && `, ${genotype.ab ? genotype.ab.toPrecision(2) : '-'}`}`}
       {variant.genotypeFilters && (
         <small>