+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## dev
 
+## 1/27/21
+* Separate structural annotation search for gCNV and genome SVs
+
+## 1/26/21
+* Improved usage for docker-compose pipeline runner
+* Add support for serving the django media root out of a google cloud storage bucket
+
+## 1/21/21
+* Support setting explicit order for saved search display (REQUIRES DB MIGRATION)
+
 ## 1/7/21
 * Update variant layout
 * Project page performance optimization